(GH-3) Refactor `fizztool` into subcommands
This change reimplements `fizztool` from an application with a single
root command that takes flags into an application with multiple
subcommands.

In this first iteration, all of the code is colocated in the `root.go`
file. In the future, the commands and utility functions should be split
out into semantically related source files so they can be independently
tested.

This change:

- Ensures the version and copyright information are correct when
  emitted, fixes #1 and fixes #2.
- Reimplements the primary functionality from the root command to the
  `get` command, which resolves #3
- Enables the automatic addition of the `help` and `completion`
  commands from cobra, previously unavailable because the application
  had no subcommands.
- Adds a `version` command that emits JSON for detailed information
  with a flag for a one-line string, which resolves #4
- Adds colorization to the JSON output, which resolves #5